+++ frontend/src/components/permissions/component.js
@@ -4,13 +4,13 @@ import { Permissions } from './permissions';
 import { Provider } from "react-redux";
 import store from "../../store";
 
-window.permissionsTable = ( elementId ) => {
+window.permissionsTable = ( elementId, is_owner ) => {
   
     $(window).on('load', function () {
         $("#content").show();
         ReactDOM.render(
             <Provider store={store}>
-                <Permissions elementId={elementId} />
+                <Permissions elementId={elementId} isOwner={is_owner} />
             </Provider>,
             document.getElementById('perm-table')
         );